feat: add reviewpolicy to manage authorization for review actions

+++ app/Http/Controllers/ReviewController.php
@@ -7,11 +7,13 @@ use App\Http\Resources\ReviewResource;
 use App\Models\Event;
 use App\Models\Review;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class ReviewController extends Controller
 {
     public function index(Request $request)
     {
+        Gate::authorize('viewAny', Review::class);
         $query = Review::query();
 
         if ($request->filled('rating')) {
@@ -24,7 +26,9 @@ class ReviewController extends Controller
 
     public function store(VerifyEventAttendanceRequest $request, Event $event)
     {
+        Gate::authorize('create', Review::class);
         $request->validateAttendance(false);
+
         $review = $event->reviews()->create([
             ...$request->validate([
                 'rating' => 'required|integer|min:1|max:5',
@@ -38,11 +42,14 @@ class ReviewController extends Controller
 
     public function show(Review $review)
     {
+        Gate::authorize('view', $review);
         return new ReviewResource($review->load(['user', 'event']));
     }
 
     public function update(Request $request, Review $review)
     {
+        Gate::authorize('update', $review);
+
         $review->update(
             $request->validate([
                 'rating' => 'sometimes|integer|min:1|max:5',
@@ -54,6 +61,7 @@ class ReviewController extends Controller
 
     public function destroy(Review $review)
     {
+        Gate::authorize('delete', $review);
         $review->delete();
         return response()->noContent();
     }
